backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
pydantic==2.8.2
pyjwt==2.9.0
psycopg2-binary==2.9.9
influxdb-client==1.43.0
paho-mqtt==1.6.1
numpy==1.26.4
scikit-learn==1.5.1
torch==2.3.0
uvicorn==0.30.1
websockets==12.0
cryptography==42.0.8
orjson==3.10.3
